requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-dotenv==1.4.2
django-revproxy==0.10.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
Pillow==7.2.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
